ensure newline at eof
